Prototype an 'abstracted' portlib generator
The idea here is that some dependent project that wished to create an
extended portlib would simply use a CMakeLists.txt that would look
roughly like the one in port/CMakeLists.txt:

    include(cmake/portlib.cmake)
    create_omr_portlib($<PROJECTNAME>)

This would then export both a static library target and an OBJECT
library target, either of which could be packaged up into the port
library by a dependent project.

This commit isn't the final word on something like this however. As it
exists today, it suffers from a number of deficiencies:

1. Today the call to create_omr_portlib must happen in port/, as the
   paths have not been correctly rooted. This is doable, but will take
   a little time to get right as we will have to deal with VPATH
   resolution.
2. This does not attempt to take any steps towards componentizing the
   build (such that the port library can build on its own, ie, running
   cmake $(OMRDIR)/port from some build directory), nor does it attempt
   to create a module for the port library, allowing something like
   FindPackage(OMR_port) to work.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>